refactor: move batch_store and batch_delete methods into FileList
