+++ lib/uploadcare/entity/file_list.rb
@@ -1,12 +1,15 @@
 # frozen_string_literal: true
 
-# Resource for FileList entity - a paginated list of files with info about whole collection
-# https://uploadcare.com/api-refs/rest-api/v0.5.0/#operation/filesList
+# Resource representing lists of files -
+# both as collection and as a paginated list of files with info about whole collection
+
+require 'uploadcare/entity/file'
 
 module Uploadcare
   class FileList < ApiStruct::Entity
     client_service FileListClient
 
-    attr_entity :next, :previous, :total, :per_page, :results
+    attr_entity :next, :previous, :total, :per_page, :results, :result, :files
+    has_entities :files, as: Uploadcare::File
   end
 end
