+++ spec/uploadcare/client/file_list_client_spec.rb
@@ -27,5 +27,49 @@ module Uploadcare
         end
       end
     end
+
+    describe 'batch_store' do
+      it 'changes files` statuses to stored' do
+        VCR.use_cassette('rest_file_batch_store') do
+          uuids = ['e9a9f291-cc52-4388-bf65-9feec1c75ff9', 'c724feac-86f7-447c-b2d6-b0ced220173d']
+          response = subject.batch_store(uuids)
+          response_value = response.value!
+          expect(uuids.all? { |uuid| response_value.to_s.include?(uuid) }).to be true
+        end
+      end
+
+      context 'invalid uuids' do
+        it 'returns a list of problems' do
+          VCR.use_cassette('rest_file_batch_store_fail') do
+            uuids = ['nonexistent', 'other_nonexistent']
+            response = subject.batch_store(uuids)
+            expect(response.success[:files]).to be_empty
+            expect(response.success[:problems]).not_to be_empty
+          end
+        end
+      end
+    end
+
+    describe 'batch_delete' do
+      it 'changes files` statuses to stored' do
+        VCR.use_cassette('rest_file_batch_delete') do
+          uuids = ['935ff093-a5cf-48c5-81cf-208511bac6e6', '63be5a6e-9b6b-454b-8aec-9136d5f83d0c']
+          response = subject.batch_delete(uuids)
+          response_value = response.value!
+          expect(response_value[:files][0][:datetime_removed]).not_to be_empty
+        end
+      end
+
+      context 'invalid uuids' do
+        it 'returns a list of problems' do
+          VCR.use_cassette('rest_file_batch_delete_fail') do
+            uuids = ['nonexistent', 'other_nonexistent']
+            response = subject.batch_delete(uuids)
+            expect(response.success[:files]).to be_empty
+            expect(response.success[:problems]).not_to be_empty
+          end
+        end
+      end
+    end
   end
 end
